refactor: CardDetailView 폴더 위치 변경
