+++ src/component/CardDetailPage/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import St from "./styled";
 import CardDetailHeader from "./CardDetailHeader";
-import CardDetailView from "./CardDetailView";
+import CardDetailView from "../CardDetailView";
 import CardDetailForm from "./CardDetailForm";
 function CardDetailPage() {
   return (
