+++ src/component/CardDetailView/CardDetailView.stories.tsx
