+++ src/component/CardDetailView/index.tsx
