+++ src/component/CardDetailView/styled.tsx
